+++ src/components/dictionaries/macmillan/View.tsx
@@ -1,4 +1,4 @@
-import React, { FC } from 'react'
+import React, { FC, ReactNode } from 'react'
 import Speaker from '@/components/Speaker'
 import StarRates from '@/components/StarRates'
 import {
@@ -8,10 +8,13 @@ import {
 } from './engine'
 import { ViewPorps } from '@/components/dictionaries/helpers'
 
-export const DictMacmillan: FC<ViewPorps<MacmillanResult>> = ({ result }) => {
+export const DictMacmillan: FC<ViewPorps<MacmillanResult>> = ({
+  result,
+  searchText
+}) => {
   switch (result.type) {
     case 'lex':
-      return renderLex(result)
+      return renderLex(result, renderSelect(result, searchText))
     case 'related':
       return renderRelated(result)
     default:
@@ -21,25 +24,54 @@ export const DictMacmillan: FC<ViewPorps<MacmillanResult>> = ({ result }) => {
 
 export default DictMacmillan
 
-function renderLex(result: MacmillanResultLex) {
-  return (
-    <>
-      {result.items.map(def => (
-        <section key={def.senses}>
-          <h1 className="dictMacmillan-Title">{def.title}</h1>
-          <header className="dictMacmillan-Header">
-            {def.ratting && <StarRates rate={def.ratting} />}
-            <span className="dictMacmillan-Header_Info">
-              {def.phsym} <Speaker src={def.pron} /> {def.pos} {def.sc}
-            </span>
-          </header>
-          <ol
-            className="dictMacmillan-Sences"
-            dangerouslySetInnerHTML={{ __html: def.senses }}
-          />
-        </section>
+function renderSelect(
+  result: MacmillanResultLex,
+  searchText: ViewPorps<MacmillanResultLex>['searchText']
+) {
+  return result.relatedEntries.length > 0 ? (
+    <select
+      style={{ width: '100%' }}
+      value={''}
+      onChange={e => {
+        if (e.currentTarget.value) {
+          searchText({
+            id: 'macmillan',
+            payload: { href: e.currentTarget.value }
+          })
+        }
+      }}
+    >
+      <option value="">
+        {result.title + (result.pos ? ` ${result.pos.toUpperCase()}` : '')}
+      </option>
+      {result.relatedEntries.map(({ title, href }) => (
+        <option key={href} value={href}>
+          {title}
+        </option>
       ))}
-    </>
+    </select>
+  ) : null
+}
+
+function renderLex(result: MacmillanResultLex, select: ReactNode) {
+  return (
+    <section onClick={onEntryClick}>
+      {select}
+      <h1 className="dictMacmillan-Title">{result.title}</h1>
+      <header className="dictMacmillan-Header">
+        {result.ratting! > 0 && <StarRates rate={result.ratting} />}
+        <span className="dictMacmillan-Header_Info">
+          {result.phsym} <Speaker src={result.pron} /> {result.pos} {result.sc}
+        </span>
+      </header>
+      <ol
+        className="dictMacmillan-Sences"
+        dangerouslySetInnerHTML={{ __html: result.senses }}
+      />
+      {result.toggleables.map((toggleable, i) => (
+        <div key={i} dangerouslySetInnerHTML={{ __html: toggleable }} />
+      ))}
+    </section>
   )
 }
 
@@ -47,10 +79,46 @@ function renderRelated(result: MacmillanResultRelated) {
   return (
     <>
       <p>Did you mean:</p>
-      <ul
-        className="dictMacmillan-Related"
-        dangerouslySetInnerHTML={{ __html: result.list }}
-      />
+      <ul className="dictMacmillan-Related">
+        {result.list.map(({ title, href }) => (
+          <li key={href}>
+            <a href={href}>{title}</a>
+          </li>
+        ))}
+      </ul>
     </>
   )
+}
+
+function onEntryClick(event: React.MouseEvent<HTMLElement>) {
+  if (!event.target['classList']) {
+    return
+  }
+  const target = event.target as Element
+
+  let isToggleHead =
+    target.classList.contains('toggle-open') ||
+    target.classList.contains('toggle-close')
+
+  if (!isToggleHead) {
+    for (let el: Element | null = target; el; el = el.parentElement) {
+      if (el.classList.contains('toggle-toggle')) {
+        isToggleHead = true
+        break
+      }
+    }
+  }
+
+  if (!isToggleHead) {
+    return
+  }
+
+  for (let el: Element | null = target; el; el = el.parentElement) {
+    if (el.classList.contains('toggleable')) {
+      el.classList.toggle('closed')
+      event.preventDefault()
+      event.stopPropagation()
+      break
+    }
+  }
 }