+++ src/content/redux/init.ts
@@ -91,12 +91,20 @@ export const init = (
         return isPopupPage() ? Promise.resolve(true) : Promise.resolve()
       }
 
+      case 'ADD_NOTEBOOK': {
+        if (msg.payload.popup === isPopupPage()) {
+          dispatch({ type: 'ADD_TO_NOTEBOOK' })
+          return Promise.resolve(true)
+        }
+        return
+      }
+
       case 'QS_PANEL_SEARCH_TEXT':
         if (isQuickSearchPage()) {
           // request searching text, from other tabs
           dispatch({ type: 'SEARCH_START', payload: { word: msg.payload } })
 
-          if (getState().isPinned) {
+          if (getState().isQSFocus) {
             // focus standalone panel
             message.send({ type: 'OPEN_QS_PANEL' })
           }