+++ src/content/redux/modules/state.ts
@@ -19,6 +19,7 @@ export const initState = () => {
       mouseY: 0,
       self: false,
       dbClick: false,
+      altKey: false,
       shiftKey: false,
       ctrlKey: false,
       metaKey: false,
@@ -32,6 +33,7 @@ export const initState = () => {
      * which could be in a standalone window or in-page element.
      */
     isQSPanel: isQuickSearchPage(),
+    isQSFocus: config.qsFocus,
     /** is a standalone quick search panel running */
     withQssaPanel: false,
     wordEditor: {
@@ -44,7 +46,7 @@ export const initState = () => {
     isShowDictPanel: isStandalonePage(),
     isExpandMtaBox: false,
     isExpandWaveformBox: false,
-    isPinned: isQuickSearchPage() ? config.qsFocus : false,
+    isPinned: false,
     /** Is current word in Notebook */
     isFav: false,
     bowlCoord: { x: 0, y: 0 },